requirements.txt
torch
wandb
matplotlib
accelerate
datasets
huggingface_hub
Pillow
torchvision
tqdm
diffusers